+++ src/workspacedefs.c
@@ -166,7 +166,7 @@ workspacedefs_clicked(GtkWidget *wid, Workspacedefs *workspacedefs)
 	g_autofree char *txt =
 		text_view_get_text(GTK_TEXT_VIEW(workspacedefs->text));
 	if (!workspacedefs_set_text(workspacedefs, txt))
-		workspace_set_show_error(workspacedefs->ws, TRUE);
+		workspace_show_error(workspacedefs->ws);
 	else
 		workspace_set_modified(workspacedefs->ws, TRUE);
 }